APPSDK-81 - added check for cookies being enabled as part of base framework APPSDK-95 - fixed nesting of widgets APPSDK-98 - fixed timezone offset parsing problem (changed from int to float) APPSDK-115 - hide widgets until compiled automatically

+++ src/java/org/appcelerator/messaging/Message.java
@@ -48,7 +48,7 @@ public class Message implements Serializable
     private MessageDataType dataType = MessageDataType.XML;
     private long timestamp;
     private long sendTimestamp;
-    private int timezoneOffset;
+    private float timezoneOffset;
     private String scope;
     private InetAddress address;
     private String version;
@@ -130,7 +130,7 @@ public class Message implements Serializable
      * @param sentTimestamp message sender timestamp
      * @param tz        timezone offset of message sender
      */
-    public Message(Subject user, String sessionid, String instanceid, String requestid, String type, MessageDirection direction, MessageDataType dataType, IMessageDataObject data, String scope, InetAddress addr, String version, long sentTimestamp, int timezoneOffset)
+    public Message(Subject user, String sessionid, String instanceid, String requestid, String type, MessageDirection direction, MessageDataType dataType, IMessageDataObject data, String scope, InetAddress addr, String version, long sentTimestamp, float timezoneOffset)
     {
         super();
         this.user = user;
@@ -390,12 +390,12 @@ public class Message implements Serializable
         this.version = version;
     }
 
-    public int getTimezoneOffset()
+    public float getTimezoneOffset()
     {
         return timezoneOffset;
     }
 
-    public void setTimezoneOffset(int timezoneOffset)
+    public void setTimezoneOffset(float timezoneOffset)
     {
         this.timezoneOffset = timezoneOffset;
     }
